requirements.txt
discord.py
yt-dlp
spotipy
async-timeout
